+++ R/shinyscan.R
@@ -4,6 +4,14 @@
 #' @export
 shinyscan <- function() {
   
+  if (!requireNamespace("scplot", quietly = TRUE)) {
+    message("You need to install the 'scplot' package to run this app ",
+            "with install.packages('scplot')")
+  }
+  if (!requireNamespace("shinyjs", quietly = TRUE)) {
+    message("You need to install the 'shinyjs' package to run this app ",
+            "with install.packages('shinyjs')")
+  }
   if (requireNamespace("shiny", quietly = TRUE)) {
     shiny::runApp(system.file('shiny', package = 'scan'),launch.browser = TRUE)
   } else {